+++ Governator-main/governator/src/app/dashboard/loading.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const loading = () => {
+  return (
+    <div className="h-screen flex items-center justify-center">
+      <div className="loading loading-dots "></div>
+    </div>
+  );
+};
+
+export default loading;
